float to double

+++ simulation/camera.h
@@ -13,10 +13,10 @@ public:
             point3 lookfrom,
             point3 lookat,
             vec3   vup,
-            double vfov, // vertical field-of-view in degrees
-            double aspect_ratio,
-            double aperture,
-            double focus_dist
+            float vfov, // vertical field-of-view in degrees
+            float aspect_ratio,
+            float aperture,
+            float focus_dist
     ) {
         auto theta = degrees_to_radians(vfov);
         auto h = tan(theta/2);
@@ -36,7 +36,7 @@ public:
     }
 
 
-    ray get_ray(double s, double t) const {
+    ray get_ray(float s, float t) const {
         vec3 rd = lens_radius * random_in_unit_disk();
         vec3 offset = u * rd.x + v * rd.y;
 
@@ -52,7 +52,7 @@ private:
     vec3 horizontal;
     vec3 vertical;
     vec3 u, v, w;
-    double lens_radius;
+    float lens_radius;
 };
 
 #endif //RAY_TRACING_CAMERA_H
